Начинаем отделять вью от обработки событий
