+++ js/templates/result-attempts-left-view.js
