+++ js/templates/result-time-left-view.js
